pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geobench-vlm"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.34.0",
    "typer",
    "torch==2.7.1",
    "torchvision==0.22.1",
    "pandas",
    "Pillow",
    "einops",
    "rich",
    "safetensors",
    "sentencepiece",
    "hf_transfer",
    "bitsandbytes>=0.42.0",

]

[dependency-groups]
lfm = ["transformers>=5.1"]
llava = ["transformers==4.52.4", "protobuf"]
llavaone1 = [
    "transformers==4.40.0",
    "llava @ git+https://github.com/LLaVA-VL/LLaVA-NeXT.git@df179663ae8b83207df100a1f7af24caec633ff9",
    "protobuf",
]
qwen = ["transformers>=4.48.0,<5.0", "qwen-vl-utils", "flash-attn>=2.7.0.post2"]
internvl = ["transformers==4.37.2", "timm==0.9.12", "flash-attn>=2.7.0.post2"]

[tool.uv]
no-build-isolation-package = ["flash-attn"]
conflicts = [
    [
        { group = "lfm" },
        { group = "llava" },
        { group = "llavaone1" },
        { group = "qwen" },
        { group = "internvl" },
    ],
]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[project.scripts]
geobench-single = "geobench_vlm.run_single:app"
geobench-temporal = "geobench_vlm.run_temporal:app"
geobench-inspect = "geobench_vlm.run_inspect:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["geobench_vlm*"]
